#9 不要になった、spot-userモデルを削除
